requirements.txt
python-telegram-bot[webhooks]==21.6
Flask==3.0.3
requests==2.32.3
paddlepaddle==2.6.0
paddleocr==2.7.0.3
Pillow==10.4.0

